fix: Use a ref for tracking position in Konami code
In React 18, state updates are no longer flushed synchronously to the
component within event handlers. Rather than use state for this purpose,
we should use a mutable ref so that we can easily increment the value
and be sure that we're always referencing the latest value and not
looking at a stale "position" number that hasn't yet received the update
from the last event that was processed.